Add RRD_LOCK flag to rrd_open()
My aim is to add read-locking to RRD.

Adding the RRD_LOCK flag will make this process straightforward.
rrd_lock() should not be needed any more, but is preserved for librrd
compat.

So `RRD_LOCK|RRD_READONLY` is intended to take a read lock, however I must
admit to limitations of this new code:

* On the Windows _locking() function, read locks are treated the same as
  write locks.

  https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/locking

* On rados, our write lock expires after 2 seconds.  This write-lock is
  probably useful as a way to detect conflicts.  However I would not want
  to rely on a lock that works like this to guarantee consistent reads for
  a backup.  Fortunately rados implements snapshots, and when we *write* to
  RRDs, we use a single atomic operation.