fadvise and madvise code in rrd_open can be kept together
There was no reason to break up fadvise and madvise.

Putting them together makes it _slightly_ clearer that the exising #ifdefs
are intended.  Documentation on the effect of fadvise on mmap() is
hard to find, so it was simpler to avoid doing that.  Also see code
comments in next commit.