fix cleanup in rrd_open()
* rados
  * don't leak rrd_file->rados
  * don't mysteriously close STDIN
* !HAVE_MMAP
  * don't leak rrd->stat_head

I have no explanation for why we were chosing to leak rrd->stat_head
(and only rrd->stat_head).  If it _was_ required for some reason, there was
a bug anyway, as we did not leak it if we fail to allocate live_hdr.

If it's required by legacy ABI, then we should remove rrd_open() from the
public API (and bump the soname).  It was originally scheduled for removal
in v1.4.