Passing RRD_READAHEAD to rrd_open() had no effect on current linux, fix it
MAP_NONBLOCK (since Linux 2.5.46)
       Only meaningful in conjunction with MAP_POPULATE.  Don't perform
       read-ahead:  create  page tables entries only for pages that are
       already present in RAM.  Since Linux 2.6.23,  this  flag  causes
       MAP_POPULATE  to  do  nothing.   One  day,  the  combination  of
       MAP_POPULATE and MAP_NONBLOCK may be reimplemented.

RRD_READAHEAD has been "broken" for a looong time.  But I don't think it's
a good idea to carry code for RRD_READAHEAD which does nothing, at the same
time as we disable the default readahead by passing MADV_RANDOM!

Solution would not necessarily be API-quality, as changing the kernel
access hint from RANDOM to SEQUENTIAL means it's allowed to do drop-behind.
I am crudely assuming the in-tree users wouldn't have a problem with this.